requirements.txt
Django==1.4.2
South==0.7.6
argparse==1.2.1
distribute==0.6.28
dj-database-url==0.2.1
django-tastypie==0.9.11
mimeparse==0.1.3
psycopg2==2.4.5
python-dateutil==1.5
wsgiref==0.1.2
